Move input-file utility functions to inputs.py
